requirements.txt
mysql-connector-python==8.0.27
pandas==1.3.5
python-dotenv==0.19.2
